- In App.xaml.cs, fetched the application user data directory from Log4net instead of SettingsHelper - Extended RingtoetsUserDataFolderConverter with a postfix path option -- Now using the option for "WTI\Ringtoets" - Various cleanups - Modified startup message Related to Issue [WTI-751]
git-svn-id: https://repos.deltares.nl/repos/WettelijkToetsInstrumentarium/trunk@4469 5a3db67b-de53-47b5-99c8-a1c30a6650e2

Former-commit-id: bbec302289cfe7c78cf9a199b74abb9267b546f1
Former-commit-id: 253ea8e50cbc202a34a11b786a27a0bcd6cc60ab

+++ Application/Ringtoets/src/Application.Ringtoets/App.xaml.cs
@@ -43,6 +43,7 @@ using Core.Plugins.CommonTools;
 using Core.Plugins.Map;
 using Core.Plugins.ProjectExplorer;
 using log4net;
+using log4net.Appender;
 using Ringtoets.GrassCoverErosionInwards.Plugin;
 using Ringtoets.HeightStructures.Plugin;
 using Ringtoets.Integration.Data;
@@ -79,7 +80,7 @@ namespace Application.Ringtoets
         {
             SetLanguage();
 
-            log.Info(Core.Common.Gui.Properties.Resources.App_Starting_Ringtoets);
+            log.Info(string.Format(Core.Common.Gui.Properties.Resources.App_Starting_Ringtoets_version_0, SettingsHelper.ApplicationVersion));
         }
 
         /// <summary>
@@ -184,13 +185,42 @@ namespace Application.Ringtoets
         /// </summary>
         private void DeleteOldLogFiles()
         {
-            string settingsDirectory = SettingsHelper.GetApplicationLocalUserSettingsDirectory();
+            string settingsDirectory = GetLogFileDirectory();
+            if (string.IsNullOrWhiteSpace(settingsDirectory))
+            {
+                return;
+            }
             foreach (string logFile in Directory.GetFiles(settingsDirectory, "*.log"))
             {
                 if ((DateTime.Now - File.GetCreationTime(logFile)).TotalDays > numberOfDaysToKeepLogFiles)
                 {
                     File.Delete(logFile);
                 }
+            }
+        }
+
+        private string GetLogFileDirectory()
+        {
+            var fileAppender = LogManager.GetAllRepositories()
+                                         .SelectMany(r => r.GetAppenders())
+                                         .OfType<FileAppender>()
+                                         .FirstOrDefault();
+            if (fileAppender == null)
+            {
+                return string.Empty;
+            }
+
+            try
+            {
+                return Path.GetDirectoryName(fileAppender.File);
+            }
+            catch (Exception exception)
+            {
+                if (exception is ArgumentException || exception is PathTooLongException)
+                {
+                    return string.Empty;
+                }
+                throw;
             }
         }
 
@@ -318,14 +348,12 @@ namespace Application.Ringtoets
                         }
                     }
                 })
-                {
                     if (exceptionDialog.ShowDialog() == DialogResult.OK)
                     {
                         Restart();
 
                         return;
                     }
-                }
             }
 
             Environment.Exit(1);
